strAsJson will sometimes have no fields
strAsJson will sometimes have no fields. This causes the plugin to crash. This is just one way to fix it. I'm mostly passing this along so the issue is known. I wish I could provide a test, but minimizing the issue would take me too long, and the code is proprietary.