Update avrohugger Scavro serialization tests
Now that sbt-avro has merged the fix for sorting schemas on different filesystems, most Scavro serialzation tests can be restored. The remaining two that cannot must wait at least until sbt-avro has a solution for duplicate schemas, because as it is now, it crashes. It should be possible to make sbt-avro handle this situation by doing what was done for avrohugger.